Move Frame database queries to /databases/:database/query
POST on the databases collection read as "create a database". The query
now lives at POST /frames/:frameId/databases/:database/query with the
database name as a validated path param and only the SQL in the body.
Authorization stays directory-scoped on the databases index so it guards
the sub-route too.
