+++ front-api/routes/v1/w/[wId]/sandbox/frames/[frameId]/databases/index.ts
@@ -1,13 +1,6 @@
-import {
-  listFrameDatabases,
-  queryFrameDatabase,
-} from "@app/lib/api/frames/databases";
+import { listFrameDatabases } from "@app/lib/api/frames/databases";
 import { canWriteFrameV2Source } from "@app/lib/api/frames/permissions";
-import { isValidPodDatabaseName } from "@app/lib/api/sandbox/db";
-import type {
-  LiveDatabaseEntry,
-  QueryDatabaseResult,
-} from "@app/lib/api/sandbox_functions/dsbx_db";
+import type { LiveDatabaseEntry } from "@app/lib/api/sandbox_functions/dsbx_db";
 import { isResourceSId } from "@app/lib/resources/string_ids";
 import type { SandboxFrameCtx } from "@front-api/middlewares/ctx";
 import { sandboxFrameApp } from "@front-api/middlewares/ctx";
@@ -19,13 +12,10 @@ import { withFrame } from "@front-api/middlewares/with_frames";
 import { createMiddleware } from "hono/factory";
 import { z } from "zod";
 
+import databaseById from "./[database]";
+
 const FrameParamsSchema = z.object({
   frameId: z.string().refine((value) => isResourceSId("file", value)),
-});
-
-const FrameDatabaseQueryRequestSchema = z.object({
-  database: z.string().refine(isValidPodDatabaseName),
-  sql: z.string().min(1),
 });
 
 type FrameDatabaseListResponse = {
@@ -34,7 +24,7 @@ type FrameDatabaseListResponse = {
 
 /**
  * @cc [owner:davidebbo,label:security;product] frame-database-authorization
- * A Frame database request MUST be granted only to callers who can write the Frame's source
+ * A Frame database request (this listing and every route under `/:database`) MUST be granted only to callers who can write the Frame's source
  * (`canWriteFrameV2Source`), with the Frame resolved inside the token's workspace, and the check
  * MUST run before the Frame sandbox is started or queried. The action-token requirement itself is
  * enforced by the parent `sandboxAuth({ allowedTokenKinds: ["action"] })` mount. Userless runs
@@ -56,7 +46,8 @@ function requireFrameDatabaseAccess() {
   });
 }
 
-// Mounted at /api/v1/w/:wId/sandbox/frames/:frameId/databases.
+// Mounted at /api/v1/w/:wId/sandbox/frames/:frameId/databases. The auth chain below is
+// directory-scoped: it also guards the /:database sub-routes.
 const app = sandboxFrameApp();
 
 app.use("*", validate("param", FrameParamsSchema));
@@ -92,38 +83,6 @@ app.get("/", async (ctx): HandlerResult<FrameDatabaseListResponse> => {
   return ctx.json({ items: result.value }, 200);
 });
 
-/**
- * @ignoreswagger
- * internal endpoint
- */
-app.post(
-  "/",
-  validate("json", FrameDatabaseQueryRequestSchema),
-  async (ctx): HandlerResult<QueryDatabaseResult> => {
-    const result = await queryFrameDatabase(
-      ctx.get("auth"),
-      ctx.get("frame"),
-      ctx.req.valid("json")
-    );
-    if (result.isErr()) {
-      const isBadQuery = result.error.code === "reconcile_blocked";
-      return apiError(
-        ctx,
-        {
-          status_code: isBadQuery ? 400 : 500,
-          api_error: {
-            type: isBadQuery
-              ? "invalid_request_error"
-              : "internal_server_error",
-            message: result.error.message,
-          },
-        },
-        result.error
-      );
-    }
-
-    return ctx.json(result.value, 200);
-  }
-);
+app.route("/:database", databaseById);
 
 export default app;
